requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.6